requirements.txt
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyjsparser==2.7.1
pytube3==9.6.4
pytz==2020.1
requests==2.24.0
six==1.15.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
